requirements/base.txt
Django==1.10.5

django-environ==0.4.1

djangorestframework==3.6.2
django-pipeline==1.6.13
django-pipeline-browserify==0.6.1
PyReact==0.1.1